Move ILValidator class into its own files
Although the ILValidator uses an IL iterator internally, it itself has
little to do with the implementation of the iterators. It is therefore
preferable to move the ILValidator implementation into its own files.

This commit also adds the new ILValidator source file to the file lists
in the makefiles of the test compiler and JitBuilder.

Lastly, the JavaILValidator class is also moved into the Java project.
